Finish registration delivery cleanup
