--- test/shared/email-renderer.test.ts
@@ -1,849 +1,0 @@
-import { expect } from "@std/expect";
-import { afterEach, beforeEach, describe, it as test } from "@std/testing/bdd";
-import { stub } from "@std/testing/mock";
-import { Liquid } from "liquidjs";
-import { ALL_SETTINGS_KEYS, settings } from "#shared/db/settings.ts";
-import type { TemplateData } from "#shared/email-renderer.ts";
-import {
-  buildTemplateData,
-  renderEmailContent,
-  renderTemplate,
-  resetEngine,
-  validateTemplate,
-} from "#shared/email-renderer.ts";
-import { describeWithEnv } from "#test-utils/db.ts";
-import { createTestGroup } from "#test-utils/db-helpers/groups.ts";
-import { createTestListing } from "#test-utils/db-helpers/listings.ts";
-import { makeTestEntry as makeEntry } from "#test-utils/factories.ts";
-import { useSetting } from "#test-utils/settings.ts";
-
-const renderConfirmation = async (): Promise<{
-  data: TemplateData;
-  result: Awaited<ReturnType<typeof renderEmailContent>>;
-}> => {
-  settings.invalidateCache();
-  await settings.loadKeys(ALL_SETTINGS_KEYS);
-  const entries = [makeEntry()];
-  const data = await buildTemplateData(
-    entries,
-    "GBP",
-    "https://example.com/t/ABC",
-  );
-  const result = await renderEmailContent("confirmation", data);
-  return { data, result };
-};
-
-describeWithEnv("email-renderer", { db: true }, () => {
-  useSetting({ currency: "GBP" });
-  beforeEach(resetEngine);
-  afterEach(resetEngine);
-
-  describe("buildTemplateData", () => {
-    test("builds correct data shape from single entry", async () => {
-      const entries = [makeEntry()];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      expect(data.listing_names).toBe("Test Listing");
-      expect(data.ticket_url).toBe("https://example.com/t/ABC");
-      expect(data.currency).toBe("GBP");
-      expect(data.entries.length).toBe(1);
-      expect(data.entries[0]!.listing.name).toBe("Test Listing");
-      expect(data.entries[0]!.listing.slug).toBe("test-listing");
-      expect(data.entries[0]!.listing.is_paid).toBe(false);
-      expect(data.attendee.name).toBe("Jane Doe");
-      expect(data.attendee.email).toBe("jane@example.com");
-    });
-
-    test("builds correct data shape from multiple entries", async () => {
-      const entries = [
-        makeEntry({ name: "Listing A" }),
-        makeEntry({ name: "Listing B" }),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC+DEF",
-      );
-
-      expect(data.listing_names).toBe("Listing A and Listing B");
-      expect(data.entries.length).toBe(2);
-      expect(data.attendee.name).toBe("Jane Doe");
-    });
-
-    test("formats three or more listing names with commas and 'and'", async () => {
-      const entries = [
-        makeEntry({ name: "Listing A" }),
-        makeEntry({ name: "Listing B" }),
-        makeEntry({ name: "Listing C" }),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC+DEF+GHI",
-      );
-
-      expect(data.listing_names).toBe("Listing A, Listing B, and Listing C");
-    });
-
-    test("marks paid listings correctly", async () => {
-      const entries = [makeEntry({ unit_price: 1000 })];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      expect(data.entries[0]!.listing.is_paid).toBe(true);
-    });
-
-    test("marks can_pay_more listings as paid", async () => {
-      const entries = [makeEntry({ can_pay_more: true, unit_price: 0 })];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      expect(data.entries[0]!.listing.is_paid).toBe(true);
-    });
-
-    test("marks a free-base entry as paid from its booking price", async () => {
-      // A package override charges a member whose base listing is free, so the
-      // booking carries a positive price_paid even though isPaidListing is false.
-      const entries = [makeEntry({ unit_price: 0 }, { price_paid: "1500" })];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      expect(data.entries[0]!.listing.is_paid).toBe(true);
-    });
-
-    test("includes attendee date when present", async () => {
-      const entries = [makeEntry({}, { date: "2026-04-15" })];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      expect(data.entries[0]!.attendee.date).toBe("2026-04-15");
-    });
-
-    // Helper for the date_range_label tests — every case follows the same
-    // makeEntry → buildTemplateData → read label flow.
-    const dateRangeLabelFor = async (
-      listing: Partial<Parameters<typeof makeEntry>[0]>,
-      attendee: Partial<Parameters<typeof makeEntry>[1]>,
-    ): Promise<string> =>
-      (
-        await buildTemplateData(
-          [makeEntry(listing, attendee)],
-          "GBP",
-          "https://example.com/t/ABC",
-        )
-      ).entries[0]!.attendee.date_range_label;
-
-    test("date_range_label: single-day daily booking formats as a date", async () => {
-      expect(
-        await dateRangeLabelFor(
-          { duration_days: 1, listing_type: "daily" },
-          { date: "2026-04-15" },
-        ),
-      ).toContain("15 April");
-    });
-
-    test("date_range_label: multi-day booking uses en dash", async () => {
-      // The label comes from the booking's stored span (end_date is exclusive),
-      // so a 3-day booking from the 15th ends (exclusive) on the 18th.
-      expect(
-        await dateRangeLabelFor(
-          { duration_days: 3, listing_type: "daily" },
-          { date: "2026-04-15", end_date: "2026-04-18" },
-        ),
-      ).toBe("15\u201317 April 2026");
-    });
-
-    test("date_range_label: empty when no booking date", async () => {
-      expect(await dateRangeLabelFor({}, { date: null })).toBe("");
-    });
-  });
-
-  describe("package grouping", () => {
-    /** A package group (Camp Kit) with a free Tent and a paid Chair member, and
-     * the email entries for booking 2 tents + 6 chairs. `hide` sets the group's
-     * hide-listings flag. */
-    const buildPackageEntries = async (hide: boolean) => {
-      const { groups } = await import("#shared/db/groups.ts");
-      const group = await createTestGroup({
-        isPackage: true,
-        name: "Camp Kit",
-      });
-      if (hide) {
-        await groups.table.update(group.id, { hidePackageListings: true });
-      }
-      const tent = await createTestListing({
-        groupId: group.id,
-        name: "Tent",
-        unitPrice: 0,
-      });
-      const chair = await createTestListing({
-        groupId: group.id,
-        name: "Chair",
-        unitPrice: 500,
-      });
-      // Grouping is by the persisted package_group_id every booking row of the
-      // order carries, so stamp the group id on each entry's attendee.
-      return [
-        makeEntry(
-          { id: tent.id, name: "Tent", unit_price: 0 },
-          { package_group_id: group.id, price_paid: "2000", quantity: 2 },
-        ),
-        makeEntry(
-          { id: chair.id, name: "Chair", unit_price: 500 },
-          { package_group_id: group.id, price_paid: "3000", quantity: 6 },
-        ),
-      ];
-    };
-
-    test("a non-hidden package heads the email with its name and lists members", async () => {
-      const entries = await buildPackageEntries(false);
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-      expect(data.listing_names).toBe("Camp Kit");
-      expect(data.entries.map((e) => e.listing.name)).toEqual([
-        "Tent",
-        "Chair",
-      ]);
-    });
-
-    test("a hidden package collapses to one row for the buyer's confirmation", async () => {
-      const entries = await buildPackageEntries(true);
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-        { hidePackageMembers: true },
-      );
-      expect(data.listing_names).toBe("Camp Kit");
-      expect(data.entries.length).toBe(1);
-      expect(data.entries[0]!.listing.name).toBe("Camp Kit");
-      // Summed across members; the paid Chair makes the bundle paid.
-      expect(data.entries[0]!.attendee.quantity).toBe(8);
-      expect(data.entries[0]!.attendee.price_paid).toBe("5000");
-      expect(data.entries[0]!.listing.is_paid).toBe(true);
-    });
-
-    test("a collapsed hidden package of free-base members is paid from prices", async () => {
-      // Every member's base listing is free, but a package override charged the
-      // buyer, so the booking carries a positive price_paid. The collapsed row
-      // must still render as paid even though no member isPaidListing.
-      const { groups } = await import("#shared/db/groups.ts");
-      const group = await createTestGroup({
-        isPackage: true,
-        name: "Free Kit",
-      });
-      await groups.table.update(group.id, { hidePackageListings: true });
-      const memberA = await createTestListing({
-        groupId: group.id,
-        name: "A",
-        unitPrice: 0,
-      });
-      const memberB = await createTestListing({
-        groupId: group.id,
-        name: "B",
-        unitPrice: 0,
-      });
-      const entries = [
-        makeEntry(
-          { id: memberA.id, name: "A", unit_price: 0 },
-          { package_group_id: group.id, price_paid: "1000", quantity: 1 },
-        ),
-        makeEntry(
-          { id: memberB.id, name: "B", unit_price: 0 },
-          { package_group_id: group.id, price_paid: "0", quantity: 1 },
-        ),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-        { hidePackageMembers: true },
-      );
-      expect(data.entries.length).toBe(1);
-      expect(data.entries[0]!.listing.is_paid).toBe(true);
-    });
-
-    test("a dated hidden package collapses keeping the widest member's range", async () => {
-      // Members share the booked start date but carry different durations;
-      // the collapsed row must keep the package-level date and the WIDEST
-      // member's range label — hiding members must not lose the booked dates.
-      const { groups } = await import("#shared/db/groups.ts");
-      const group = await createTestGroup({
-        isPackage: true,
-        name: "Dated Kit",
-      });
-      await groups.table.update(group.id, { hidePackageListings: true });
-      const spans: [name: string, endDate: string | null][] = [
-        // A dated single-day member with no end_date sorts below any ranged stay.
-        ["Day Only", null],
-        ["Narrow A", "2026-08-02"],
-        ["Wide", "2026-08-04"],
-        ["Narrow B", "2026-08-02"],
-      ];
-      const entries = [];
-      for (const [name, endDate] of spans) {
-        const listing = await createTestListing({
-          groupId: group.id,
-          name,
-          unitPrice: 500,
-        });
-        entries.push(
-          makeEntry(
-            { id: listing.id, name, unit_price: 500 },
-            {
-              date: "2026-08-01",
-              end_date: endDate,
-              package_group_id: group.id,
-              price_paid: "500",
-              quantity: 1,
-            },
-          ),
-        );
-      }
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-        { hidePackageMembers: true },
-      );
-      expect(data.entries.length).toBe(1);
-      expect(data.entries[0]!.attendee.date).toBe("2026-08-01");
-      // end_date is exclusive, so the 3-night Wide member spans 1–3 August.
-      expect(data.entries[0]!.attendee.date_range_label).toBe(
-        "1–3 August 2026",
-      );
-    });
-
-    test("a hidden package still shows members in the admin notification", async () => {
-      const entries = await buildPackageEntries(true);
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-      expect(data.entries.map((e) => e.listing.name)).toEqual([
-        "Tent",
-        "Chair",
-      ]);
-    });
-
-    test("a mixed order conceals its hidden bundle beside plain rows", async () => {
-      // The flagship cart shape: a hidden package booked ALONGSIDE an ordinary
-      // listing. The buyer's confirmation must collapse the bundle's rows
-      // behind the package name while the plain row renders normally — a
-      // single-shared-package rule would leak every member name here.
-      const entries = await buildPackageEntries(true);
-      const lantern = await createTestListing({
-        name: "Lantern",
-        unitPrice: 300,
-      });
-      entries.push(
-        makeEntry(
-          { id: lantern.id, name: "Lantern", unit_price: 300 },
-          { price_paid: "300", quantity: 1 },
-        ),
-      );
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-        { hidePackageMembers: true },
-      );
-      expect(data.listing_names).toBe("Camp Kit and Lantern");
-      expect(data.entries.map((e) => e.listing.name)).toEqual([
-        "Camp Kit",
-        "Lantern",
-      ]);
-      // The bundle row still sums its members; the plain row keeps its own.
-      expect(data.entries[0]!.attendee.quantity).toBe(8);
-      expect(data.entries[1]!.attendee.quantity).toBe(1);
-    });
-
-    test("a standalone booking of a hidden one-member package's listing is NOT collapsed", async () => {
-      // Regression for the membership-equality bug: a one-member HIDDEN package
-      // whose sole listing is booked standalone (NOT via the package, so its
-      // booking rows carry package_group_id 0) must render normally — never
-      // renamed to or collapsed as the package — because grouping is by the
-      // persisted id, not by membership equality.
-      const { groups } = await import("#shared/db/groups.ts");
-      const group = await createTestGroup({
-        isPackage: true,
-        name: "Solo Kit",
-      });
-      await groups.table.update(group.id, { hidePackageListings: true });
-      const widget = await createTestListing({
-        groupId: group.id,
-        name: "Widget",
-        unitPrice: 500,
-      });
-      const entries = [
-        makeEntry(
-          { id: widget.id, name: "Widget", unit_price: 500 },
-          // package_group_id defaults to 0 — a standalone, non-package order.
-          { price_paid: "1500", quantity: 3 },
-        ),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-        { hidePackageMembers: true },
-      );
-      // Not collapsed, not renamed to the package.
-      expect(data.listing_names).toBe("Widget");
-      expect(data.entries.length).toBe(1);
-      expect(data.entries[0]!.listing.name).toBe("Widget");
-    });
-  });
-
-  describe("renderTemplate", () => {
-    const sampleData: TemplateData = {
-      amount_owed: "0",
-      attendee: {
-        address: "123 St",
-        date: null,
-        date_range_label: "",
-        email: "jane@test.com",
-        name: "Jane",
-        phone: "555",
-        price_paid: "2000",
-        quantity: 2,
-        special_instructions: "",
-      },
-      currency: "GBP",
-      entries: [
-        {
-          attendee: {
-            address: "123 St",
-            date: null,
-            date_range_label: "",
-            email: "jane@test.com",
-            name: "Jane",
-            phone: "555",
-            price_paid: "2000",
-            quantity: 2,
-            special_instructions: "",
-          },
-          listing: { is_paid: true, name: "Concert", slug: "concert" },
-        },
-      ],
-      listing_names: "Concert",
-      ticket_url: "https://example.com/t/ABC",
-    };
-
-    test("renders simple variable interpolation", async () => {
-      const result = await renderTemplate(
-        "Hello {{ attendee.name }}",
-        sampleData,
-      );
-      expect(result).toBe("Hello Jane");
-    });
-
-    test("renders listing_names variable", async () => {
-      const result = await renderTemplate(
-        "For {{ listing_names }}",
-        sampleData,
-      );
-      expect(result).toBe("For Concert");
-    });
-
-    test("renders ticket_url variable", async () => {
-      const result = await renderTemplate("{{ ticket_url }}", sampleData);
-      expect(result).toBe("https://example.com/t/ABC");
-    });
-
-    test("renders currency filter", async () => {
-      const result = await renderTemplate("{{ 2000 | currency }}", sampleData);
-      expect(result).toBe("£20");
-    });
-
-    test("renders currency filter with string value", async () => {
-      const result = await renderTemplate(
-        "{% for entry in entries %}{{ entry.attendee.price_paid | currency }}{% endfor %}",
-        sampleData,
-      );
-      expect(result).toBe("£20");
-    });
-
-    test("renders pluralize filter for singular", async () => {
-      const data = {
-        ...sampleData,
-        entries: [
-          {
-            ...sampleData.entries[0]!,
-            attendee: { ...sampleData.entries[0]!.attendee, quantity: 1 },
-          },
-        ],
-      };
-      const result = await renderTemplate(
-        '{% for entry in entries %}{{ entry.attendee.quantity }} {{ entry.attendee.quantity | pluralize: "ticket", "tickets" }}{% endfor %}',
-        data,
-      );
-      expect(result).toBe("1 ticket");
-    });
-
-    test("renders pluralize filter for plural", async () => {
-      const result = await renderTemplate(
-        '{% for entry in entries %}{{ entry.attendee.quantity }} {{ entry.attendee.quantity | pluralize: "ticket", "tickets" }}{% endfor %}',
-        sampleData,
-      );
-      expect(result).toBe("2 tickets");
-    });
-
-    test("renders for loop over entries", async () => {
-      const data: TemplateData = {
-        ...sampleData,
-        entries: [
-          {
-            ...sampleData.entries[0]!,
-            listing: { is_paid: false, name: "Listing A", slug: "a" },
-          },
-          {
-            ...sampleData.entries[0]!,
-            listing: { is_paid: false, name: "Listing B", slug: "b" },
-          },
-        ],
-      };
-      const result = await renderTemplate(
-        "{% for entry in entries %}{{ entry.listing.name }} {% endfor %}",
-        data,
-      );
-      expect(result).toBe("Listing A Listing B");
-    });
-
-    test("renders conditional on is_paid", async () => {
-      const result = await renderTemplate(
-        "{% for entry in entries %}{% if entry.listing.is_paid %}paid{% else %}free{% endif %}{% endfor %}",
-        sampleData,
-      );
-      expect(result).toBe("paid");
-    });
-
-    test("renders conditional on attendee date", async () => {
-      const data: TemplateData = {
-        ...sampleData,
-        entries: [
-          {
-            ...sampleData.entries[0]!,
-            attendee: {
-              ...sampleData.entries[0]!.attendee,
-              date: "2026-04-15",
-            },
-          },
-        ],
-      };
-      const result = await renderTemplate(
-        "{% for entry in entries %}{% if entry.attendee.date %}{{ entry.attendee.date }}{% endif %}{% endfor %}",
-        data,
-      );
-      expect(result).toBe("2026-04-15");
-    });
-
-    test("trims whitespace from rendered output", async () => {
-      const result = await renderTemplate("  hello  ", sampleData);
-      expect(result).toBe("hello");
-    });
-  });
-
-  describe("renderEmailContent", () => {
-    test("uses default templates when no custom templates are set", async () => {
-      const { result } = await renderConfirmation();
-
-      expect(result.subject).toContain("Test Listing");
-      expect(result.html).toContain("Test Listing");
-      expect(result.html).toContain("https://example.com/t/ABC");
-      expect(result.text).toContain("Test Listing");
-      expect(result.text).toContain("https://example.com/t/ABC");
-    });
-
-    test("uses custom templates when set", async () => {
-      await settings.update.email.template(
-        "confirmation",
-        "subject",
-        "Custom: {{ listing_names }}",
-      );
-      await settings.update.email.template(
-        "confirmation",
-        "html",
-        "<b>Custom HTML for {{ attendee.name }}</b>",
-      );
-      await settings.update.email.template(
-        "confirmation",
-        "text",
-        "Custom text for {{ attendee.name }}",
-      );
-      const { result } = await renderConfirmation();
-
-      expect(result.subject).toBe("Custom: Test Listing");
-      expect(result.html).toBe("<b>Custom HTML for Jane Doe</b>");
-      expect(result.text).toBe("Custom text for Jane Doe");
-    });
-
-    test("falls back to default on custom template render error", async () => {
-      await settings.update.email.template(
-        "confirmation",
-        "subject",
-        "{{ invalid | nonexistent_filter }}",
-      );
-      const { result } = await renderConfirmation();
-
-      expect(result.subject).toContain("Test Listing");
-      expect(result.errors).toHaveLength(1);
-      expect(result.errors[0]).toBeInstanceOf(Error);
-    });
-
-    test("renders admin notification defaults correctly", async () => {
-      const entries = [makeEntry()];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("admin", data);
-
-      expect(result.subject).toContain("Jane Doe");
-      expect(result.subject).toContain("Test Listing");
-      expect(result.html).toContain("Jane Doe");
-      expect(result.text).toContain("Name: Jane Doe");
-    });
-
-    test("renders admin notification with contact details", async () => {
-      const entries = [
-        makeEntry(
-          {},
-          {
-            address: "123 Main St",
-            phone: "555-1234",
-            special_instructions: "Wheelchair",
-          },
-        ),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("admin", data);
-
-      expect(result.text).toContain("Phone: 555-1234");
-      expect(result.text).toContain("Address: 123 Main St");
-      expect(result.text).toContain("Notes: Wheelchair");
-    });
-
-    test("omits empty contact fields in admin notification", async () => {
-      const entries = [
-        makeEntry({}, { address: "", phone: "", special_instructions: "" }),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("admin", data);
-
-      expect(result.text).not.toContain("Phone:");
-      expect(result.text).not.toContain("Address:");
-      expect(result.text).not.toContain("Notes:");
-    });
-
-    test("renders paid listing with currency in confirmation", async () => {
-      const entries = [
-        makeEntry({ unit_price: 1000 }, { price_paid: "2000", quantity: 2 }),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("confirmation", data);
-
-      expect(result.html).toContain("£20");
-      expect(result.text).toContain("£20");
-    });
-
-    test("shows the amount owed when a balance is outstanding", async () => {
-      // A provider-less booking: nothing collected, the full value owed.
-      const entries = [
-        makeEntry(
-          { unit_price: 1000 },
-          { price_paid: "0", remaining_balance: 2000 },
-        ),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const confirmation = await renderEmailContent("confirmation", data);
-      const admin = await renderEmailContent("admin", data);
-
-      expect(confirmation.html).toContain("Amount owed");
-      expect(confirmation.html).toContain("£20");
-      expect(confirmation.text).toContain("Amount owed: £20");
-      expect(admin.html).toContain("Amount owed");
-    });
-
-    test("omits the amount owed when the booking is fully paid", async () => {
-      const entries = [
-        makeEntry(
-          { unit_price: 1000 },
-          { price_paid: "1000", remaining_balance: 0 },
-        ),
-      ];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("confirmation", data);
-
-      expect(result.html).not.toContain("Amount owed");
-      expect(result.text).not.toContain("Amount owed");
-    });
-
-    test("shows date when attendee has date", async () => {
-      const entries = [makeEntry({}, { date: "2026-07-15" })];
-      const data = await buildTemplateData(
-        entries,
-        "GBP",
-        "https://example.com/t/ABC",
-      );
-
-      const result = await renderEmailContent("confirmation", data);
-
-      expect(result.html).toContain("2026-07-15");
-      expect(result.text).toContain("2026-07-15");
-    });
-
-    test("uses mix of custom and default parts", async () => {
-      await settings.update.email.template(
-        "confirmation",
-        "subject",
-        "Custom Subject: {{ listing_names }}",
-      );
-      // html and text remain default
-      const { result } = await renderConfirmation();
-
-      expect(result.subject).toBe("Custom Subject: Test Listing");
-      // html and text should still use defaults
-      expect(result.html).toContain("Thanks for registering!");
-      expect(result.text).toContain("Thanks for registering!");
-    });
-
-    test("collects non-Error thrown values", async () => {
-      await settings.update.email.template(
-        "confirmation",
-        "subject",
-        "Custom {{ listing_names }}",
-      );
-      settings.invalidateCache();
-      await settings.loadKeys(ALL_SETTINGS_KEYS);
-
-      // Stub parseAndRender to throw a non-Error value on first call, then succeed
-      let callCount = 0;
-      const original = Liquid.prototype.parseAndRender;
-      const parseAndRenderStub = stub(
-        Liquid.prototype,
-        "parseAndRender",
-        function (
-          this: InstanceType<typeof Liquid>,
-          ...args: Parameters<typeof original>
-        ) {
-          if (callCount++ === 0) throw "string error value";
-          return original.apply(this, args);
-        },
-      );
-      try {
-        const entries = [makeEntry()];
-        const data = await buildTemplateData(
-          entries,
-          "GBP",
-          "https://example.com/t/ABC",
-        );
-        const result = await renderEmailContent("confirmation", data);
-
-        expect(result.errors).toEqual(["string error value"]);
-      } finally {
-        parseAndRenderStub.restore();
-      }
-    });
-
-    test("resets to defaults after clearing custom template", async () => {
-      await settings.update.email.template(
-        "confirmation",
-        "subject",
-        "Custom Subject",
-      );
-      settings.invalidateCache();
-      await settings.loadKeys(ALL_SETTINGS_KEYS);
-
-      await settings.update.email.template("confirmation", "subject", "");
-      const { result } = await renderConfirmation();
-
-      expect(result.subject).toContain("Test Listing");
-    });
-  });
-
-  describe("validateTemplate", () => {
-    test("returns null for valid template", () => {
-      expect(validateTemplate("Hello {{ name }}")).toBeNull();
-    });
-
-    test("returns null for template with for loop", () => {
-      expect(
-        validateTemplate("{% for x in items %}{{ x }}{% endfor %}"),
-      ).toBeNull();
-    });
-
-    test("returns error for unclosed tag", () => {
-      const error = validateTemplate("{% for x in items %}{{ x }}");
-      expect(error).not.toBeNull();
-    });
-
-    test("returns error for invalid syntax", () => {
-      const error = validateTemplate("{% invalid_tag %}");
-      expect(error).not.toBeNull();
-    });
-
-    test("returns null for empty template", () => {
-      expect(validateTemplate("")).toBeNull();
-    });
-
-    test("returns string representation of non-Error thrown value", () => {
-      const parseStub = stub(Liquid.prototype, "parse", () => {
-        throw "raw string parse error";
-      });
-      try {
-        const result = validateTemplate("anything");
-        expect(result).toBe("raw string parse error");
-      } finally {
-        parseStub.restore();
-      }
-    });
-  });
-});
